feat: add new recommend_summary_agent with related utilities and refactor agent imports for modularity
