+++ agents/matmaster_agent/core_agents/public_agents/job_agents/result_core_agent/prompt.py
