+++ agents/matmaster_agent/sub_agents/DPACalculator_agent/agent.py
@@ -3,7 +3,9 @@ from google.adk.agents import BaseAgent
 from google.adk.tools.mcp_tool.mcp_session_manager import StreamableHTTPServerParams
 
 from agents.matmaster_agent.constant import MATMASTER_AGENT_NAME
-from agents.matmaster_agent.job_agents.agent import BaseAsyncJobAgent
+from agents.matmaster_agent.core_agents.public_agents.job_agents.agent import (
+    BaseAsyncJobAgent,
+)
 from agents.matmaster_agent.llm_config import LLMConfig
 from agents.matmaster_agent.logger import matmodeler_logging_handler
 from agents.matmaster_agent.sub_agents.DPACalculator_agent.callback import (
