+++ agents/matmaster_agent/sub_agents/MrDice_agent/openlam_agent/agent.py
@@ -4,7 +4,7 @@ from google.adk.agents import BaseAgent
 from google.adk.tools.mcp_tool.mcp_session_manager import SseServerParams
 
 from agents.matmaster_agent.constant import LOCAL_EXECUTOR, BohriumStorge
-from agents.matmaster_agent.core_agents.worker_agents.sync_agent import (
+from agents.matmaster_agent.core_agents.public_agents.sync_agent import (
     BaseSyncAgentWithToolValidator,
 )
 from agents.matmaster_agent.sub_agents.MrDice_agent.constant import MrDice_Agent_Name
