+++ agents/matmaster_agent/sub_agents/ScienceNavigator_agent/agent.py
@@ -6,7 +6,7 @@ from google.adk.tools.mcp_tool.mcp_session_manager import (
 )
 
 from agents.matmaster_agent.constant import MATMASTER_AGENT_NAME
-from agents.matmaster_agent.core_agents.worker_agents.sync_agent import (
+from agents.matmaster_agent.core_agents.public_agents.sync_agent import (
     BaseSyncAgentWithToolValidator,
 )
 from agents.matmaster_agent.llm_config import LLMConfig
