+++ agents/matmaster_agent/sub_agents/built_in_agent/file_parse_agent/agent.py
@@ -1,4 +1,4 @@
-from agents.matmaster_agent.core_agents.worker_agents.sync_agent import (
+from agents.matmaster_agent.core_agents.public_agents.sync_agent import (
     BaseSyncAgentWithToolValidator,
 )
 from agents.matmaster_agent.llm_config import LLMConfig
