+++ agents/matmaster_agent/sub_agents/document_parser_agent/agent.py
@@ -5,7 +5,7 @@ from google.adk.tools.mcp_tool.mcp_session_manager import StreamableHTTPServerPa
 from agents.matmaster_agent.constant import MATMASTER_AGENT_NAME, BohriumStorge
 from agents.matmaster_agent.llm_config import LLMConfig
 
-from ...core_agents.worker_agents.sync_agent import BaseSyncAgentWithToolValidator
+from ...core_agents.public_agents.sync_agent import BaseSyncAgentWithToolValidator
 from .callback import validate_document_url
 from .constant import DocumentParserAgentName, DocumentParserServerUrl
 from .prompt import DocumentParserAgentDescription, DocumentParserAgentInstruction
