+++ agents/matmaster_agent/sub_agents/traj_analysis_agent/agent.py
@@ -3,7 +3,7 @@ from google.adk.agents import BaseAgent
 from google.adk.tools.mcp_tool.mcp_session_manager import SseServerParams
 
 from agents.matmaster_agent.constant import BohriumStorge
-from agents.matmaster_agent.core_agents.worker_agents.sync_agent import (
+from agents.matmaster_agent.core_agents.public_agents.sync_agent import (
     BaseSyncAgentWithToolValidator,
 )
 from agents.matmaster_agent.llm_config import LLMConfig
