+++ agents/matmaster_agent/sub_agents/vaspkit_agent/agent.py
@@ -7,7 +7,7 @@ from agents.matmaster_agent.constant import (
     MATMASTER_AGENT_NAME,
     BohriumStorge,
 )
-from agents.matmaster_agent.core_agents.worker_agents.sync_agent import (
+from agents.matmaster_agent.core_agents.public_agents.sync_agent import (
     BaseSyncAgentWithToolValidator,
 )
 from agents.matmaster_agent.llm_config import LLMConfig
